moved SG to separate class round working - other shapes not returning correct carat weight

+++ Gemwest/CCalculator.h
@@ -6,9 +6,9 @@ using namespace System::Diagnostics;
 ref class CCalculator abstract
 {
 public:
-	//virtual Double term() const = 0;
-	virtual Double term()  = 0;
-	virtual Double recut_weight() = 0;
+	//virtual Decimal term() const = 0;
+	virtual Decimal term()  = 0;
+	virtual Decimal recut_weight() = 0;
 	//virtual String^ print() {
 	//	return System::Convert::ToString(5+"many ct");
 	//}
